Simple register

+++ app/views.py
@@ -32,6 +32,9 @@ def notes(request):
 
 
 def notes_detail(request, id):
+    if not request.user.is_authenticated:
+        return redirect("authentication:login")
+
     note = models.Note.objects.get(id=id)
     note_schema = serializers.NoteSchema()
 
@@ -46,6 +49,9 @@ def notes_detail(request, id):
 
 
 def notes_edit(request, id):
+    if not request.user.is_authenticated:
+        return redirect("authentication:login")
+
     note = models.Note.objects.get(id=id)
     note_schema = serializers.NoteSchema()
 
@@ -70,6 +76,9 @@ def notes_edit(request, id):
 
 
 def notes_delete(request, id):
+    if not request.user.is_authenticated:
+        return redirect("authentication:login")
+
     note = models.Note.objects.get(id=id)
     try:
         note.delete()
@@ -81,6 +90,9 @@ def notes_delete(request, id):
 
 
 def notes_create(request):
+    if not request.user.is_authenticated:
+        return redirect("authentication:login")
+
     if request.method == 'POST':
         try:
             note_schema = serializers.NoteSchema()
